doc: junit5 java examples and reorg (#1281)

+++ webtau-junit5-examples/src/main/java/com/example/tests/junit5/PriceCalculator.java
@@ -17,7 +17,7 @@
 package com.example.tests.junit5;
 
 public class PriceCalculator {
-    public static long calculate(long price, long quantity) {
+    public static double calculate(double price, long quantity) {
         return price * quantity;
     }
 }
